// GlobePayHub/src/app/enums.ts
export enum CardType {
  Debit = 'DEBIT',
  Credit = 'CREDIT',
}

export enum Currency {
  INR = 'INR',
  USD = 'USD',
  EUR = 'EUR',
}

export enum UserRole {
  User = 'ROLE_USER',
  Moderator = 'ROLE_MODERATOR',
  Admin = 'ROLE_ADMIN',
  Merchant = 'ROLE_MERCHANT',
  Support = 'ROLE_SUPPORT',
  Auditor = 'ROLE_AUDITOR',
  Finance = 'ROLE_FINANCE',
}